2 matches

+++ src/data/sbss/dword_800ABBDC.c
@@ -1,4 +1,4 @@
 #include "linker.h"
 
-int SECTION(".sbss") dword_800ABBDC;
+short SECTION(".sbss") dword_800ABBDC;
